refactor: split layout.tsx into app-sidebar and groups-section components
Extract AppSidebar and GroupsSection (with SortableGroupItem and
DragHandle) into their own files. layout.tsx drops from ~460 lines
to ~25 lines with a single responsibility: the layout shell.

Closes #356
